Commit vs2017RC project files.
Fixed a C4477 warning. format string '%6b' requires an argument of type
'int' but variadic argument 1 has type 'byte *'.

+++ utils/studiomdl/write.c
@@ -477,7 +477,7 @@ void WriteModel( )
 		{
 			VectorCopy( model[i]->normal[normimap[j]].org, pnorm[j] );
 		}
-		printf("vertices  %6d bytes (%d vertices, %d normals)\n", pData - cur, model[i]->numverts, model[i]->numnorms);
+		printf("vertices  %6d bytes (%d vertices, %d normals)\n", (int)(pData - cur), model[i]->numverts, model[i]->numnorms);
 		cur = (int)pData;
 
 		// save mesh info
@@ -514,7 +514,7 @@ void WriteModel( )
 			total_tris += pmesh[j].numtris;
 			total_strips += numcommandnodes;
 		}
-		printf("mesh      %6d bytes (%d tris, %d strips)\n", pData - cur, total_tris, total_strips);
+		printf("mesh      %6d bytes (%d tris, %d strips)\n", (int)(pData - cur), total_tris, total_strips);
 		cur = (int)pData;
 	}	
 }
